Use improved gzip compressor.
This pure-Go library has optimizaed gzip function. Typically this gives more ~70% for throughput for web content, as well as fewer allocations. It also changes the default (unset) compression level to 4, which in general is a better tradeoff for web content.

It also adds a "ConstantCompression" (-2) level, which is ~4x faster than *BestSpeed*, but with output typically twice the size.

For performance comparisons see the package and https://blog.klauspost.com/gzip-performance-for-go-webservers/